+++ src/utils/publishCustomFunctions.ts
@@ -7,7 +7,6 @@ import fs from 'fs-extra';
 import ora from 'ora';
 import os from 'os';
 import path from 'path';
-import prompts from 'prompts';
 import vm from 'vm';
 
 /* internal dependencies */
@@ -15,20 +14,16 @@ import vm from 'vm';
 import IDE from './ide';
 import acquireCustomFunctionsProject from './acquireCustomFunctionsProject';
 
+import {
+  MetaData,
+  NamedObject,
+  resolveMissingFunction,
+} from './publishFunctions';
+
 /* execute command */
 
 const workingDir = process.cwd();
 let identifier: string;
-
-type NamedObject = Record<string, string | object>;
-
-type MetaData = {
-  [name: string]: {
-    replace?: string;
-    returnType: string;
-    inputVariables: NamedObject;
-  };
-};
 
 type CustomFunction = {
   id: string;
@@ -45,96 +40,6 @@ type Action = {
 };
 
 type Actions = Action[];
-
-const resolveMissingFunction = async (
-  groomed: MetaData,
-  metaData: MetaData,
-  name: string,
-  body: string,
-): Promise<MetaData> => {
-  const { replace } = await prompts({
-    type: 'toggle',
-    name: 'replace',
-    message: `Function "${name}" is missing. What do you want to do?`,
-    initial: false,
-    active: 'replace',
-    inactive: 'add',
-  });
-
-  if (replace) {
-    const choices = Object.keys(metaData).filter(key => !groomed[key]);
-    let replacedFunction;
-
-    if (choices.length === 1) {
-      const confirm = await prompts({
-        type: 'confirm',
-        name: 'value',
-        message: `Replace "${choices[0]}"?`,
-        initial: true,
-      });
-      if (confirm.value) [replacedFunction] = choices;
-      else throw new Error('Abort.');
-    } else {
-      replacedFunction = (await prompts({
-        type: 'select',
-        name: 'value',
-        message: 'Replace',
-        choices: choices.map(key => ({ title: key, value: key })),
-        initial: 0,
-      })).value;
-    }
-
-    // eslint-disable-next-line no-param-reassign
-    groomed[name] = metaData[replacedFunction];
-    // eslint-disable-next-line no-param-reassign
-    groomed[name].replace = replacedFunction;
-    // eslint-disable-next-line no-param-reassign
-    delete metaData[replacedFunction];
-  } else {
-    const matches = body.match(/await context\((["'])(.*?)\1\)/g);
-    const defaultInputVariables = (matches || [])
-      .map(m => `${m.slice(15, -2)}:string`)
-      .join(' ');
-
-    const { returnType, inputVariables } = await prompts([
-      {
-        type: 'select',
-        name: 'returnType',
-        message: `What is the return type of the function?`,
-        choices: [
-          { title: 'string', value: 'string' },
-          { title: 'integer', value: 'integer' },
-          { title: 'boolean', value: 'boolean' },
-        ],
-        initial: 0,
-      } as prompts.PromptObject,
-      {
-        type: 'text',
-        name: 'inputVariables',
-        message: 'What are the input variables? (`name:type name:type ...`)',
-        initial: defaultInputVariables,
-      } as prompts.PromptObject,
-    ]);
-
-    // eslint-disable-next-line no-param-reassign
-    groomed[name] = {
-      returnType,
-      inputVariables: inputVariables.split(/(\s|,)/).reduce(
-        (variables: NamedObject, variable: string): NamedObject => {
-          if (variable.length) {
-            const [varName, varType] = variable.split(':');
-            // eslint-disable-next-line no-param-reassign
-            variables[varName] = varType;
-          }
-          return variables;
-        },
-        {} as NamedObject,
-      ),
-    };
-  }
-
-  return groomed;
-};
 
 const groomMetaData = async (): Promise<MetaData> => {
   console.log('Grooming functions.json ...');
@@ -159,12 +64,20 @@ const groomMetaData = async (): Promise<MetaData> => {
         groomed[name] = metaData[name];
 
         if (!groomed[name]) {
+          const matches = customFunctions[name]
+            .toString()
+            .match(/await context\((["'])(.*?)\1\)/g);
+
+          const defaultInputVariables = (matches || [])
+            .map(m => `${m.slice(15, -2)}:string`)
+            .join(' ');
+
           // eslint-disable-next-line no-param-reassign
           groomed = await resolveMissingFunction(
             groomed,
             metaData,
             name,
-            customFunctions[name].toString(),
+            defaultInputVariables,
           );
         }
